# server/src/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Donggle Swagger API!",
    "version": "1.0.0",
    "title": "Donggle",
    "contact": {
      "email": "notready@donggle"
    }
  },
  "host": "boost.boost-donggle.kro.kr",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
    "/auth/signin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "동글 서비스 로그인",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Sign in Objects",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "로그인 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "properties": {
                    "oauthEmail": {
                      "type": "object"
                    },
                    "address": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 수행 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다!"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류 발생 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "재로그인이 필요합니다."
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "동글 서비스 회원가입",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Sing up Objects",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "oauth_email": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "center": {
                  "type": "array",
                  "example": [
                    0,
                    0
                  ],
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "image": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "회원 가입 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "example": "성공적으로 회원가입 되었습니다."
                }
              }
            }
          },
          "500": {
            "description": "서버 오류 발생 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "회원가입 중 오류가 발생했습니다. 다시 시도해 주세요."
                }
              }
            }
          }
        }
      }
    },
    "/auth/info": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "회원 정보 요청",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "회원 정보 성공적으로 반환 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "properties": {
                    "oauthEmail": {
                      "type": "object"
                    },
                    "address": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string"
                    }
                  }
                },
                "message": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "401": {
            "description": "회원 가입 되어 있지 않은 경우 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "500": {
            "description": "회원 정보 반환 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "재로그인이 필요합니다."
                }
              }
            }
          }
        }
      }
    },
...
    "/review": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "특정 지역의 후기 데이터 요청",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "Map 지역 이름",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "Pagination 번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "itemNum",
            "in": "query",
            "description": "Fetch할 review 개수",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "후기 정보 반환 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/Review"
                },
                "message": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다."
                }
              }
            }
          },
          "500": {
            "description": "후기 정보 반환 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "후기 정보를 받아오지 못했습니다."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "review"
        ],
        "summary": "특정 지역 후기 데이터 제출 요청",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Sing up Objects",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReviewInsert"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "후기 정보 전달 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "후기를 정상적으로 저장하였습니다."
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다."
                }
              }
            }
          },
          "500": {
            "description": "후기 정보 반환 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "오류가 발생하여 후기를 저장하지 못했습니다."
                }
              }
            }
          }
        }
      }
    },
    "/review/user": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "특정 유저가 작성한 후기 데이터 요청",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "사용자 Email 정보",
            "required": true,
            "type": "string"
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "Pagination 번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "itemNum",
            "in": "query",
            "description": "Fetch할 review 개수",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "후기 정보 반환 성공 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/Review"
                },
                "message": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "400": {
            "description": "비정상적인 요청 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "잘못된 요청입니다!"
                }
              }
            }
          },
          "500": {
            "description": "후기 정보 반환 실패 시 반환",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "object",
                  "example": "{}"
                },
                "message": {
                  "type": "string",
                  "example": "후기 정보를 받아오지 못했습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Categories": {
      "type": "object",
      "properties": {
        "safety": {
          "type": "integer",
          "format": "int64"
        },
        "traffic": {
          "type": "integer",
          "format": "int64"
        },
        "food": {
          "type": "integer",
          "format": "int64"
        },
        "entertainment": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "Review": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "oauth_email": {
          "type": "string"
        },
        "categories": {
          "$ref": "#/definitions/Categories"
        },
        "hastags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ReviewInsert": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "oauth_email": {
          "type": "string",
          "format": "int64"
        },
        "categories": {
          "$ref": "#/definitions/Categories"
        },
        "hastags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}